+++ CHANGELOG.md
@@ -7,9 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/Gabwasnt/NeoVelocity/compare/dev...HEAD)
 
++ Fixed incompatibility
+  with [Forgified Fabric Networking API (v1)](https://modrinth.com/mod/forgified-fabric-api). [#10](https://github.com/Gabwasnt/NeoVelocity/issues/10)
++ **Config Structure Updated**
+    - Added new `forwarding` and `compatibility` sections.
+    - Introduced `login-custom-packet-catchall` option.
+
 ## [1.2.1](https://github.com/Gabwasnt/NeoVelocity/compare/dev...v1.2.1) - 2025-04-18
 
-+ Added support for loading Velocity forwarding secret from an external UTF-8 encoded file via 'forwarding-secret-type'. [#11](https://github.com/Gabwasnt/NeoVelocity/issues/11)
++ Added support for loading Velocity forwarding secret from an external UTF-8 encoded file via
+  'forwarding-secret-type'. [#11](https://github.com/Gabwasnt/NeoVelocity/issues/11)
 + 1.21.5 compatibility!
 
 ## [1.2.0](https://github.com/Gabwasnt/NeoVelocity/compare/dev...v1.2.0) - 2025-02-13